Refactor password service to own file
