+++ full-service/src/service/mod.rs
@@ -3,9 +3,11 @@
 //! Implementations of services.
 
 pub mod decorated_types;
+mod password_manager;
 pub mod sync;
 pub mod transaction_builder;
 pub mod wallet;
 mod wallet_impl;
 
+pub use password_manager::PasswordServiceError;
 pub use wallet_impl::WalletService;
